+++ src/components/FullscreenPlayer.tsx
@@ -13,15 +13,13 @@ import { useCachedUrl } from './CachedImage';
 import { getCachedBlob } from '../utils/imageCache';
 import { extractCoverColors } from '../utils/dynamicColors';
 import { useTranslation } from 'react-i18next';
-import { useLyrics, type WordLyricsLine } from '../hooks/useLyrics';
 import { useAuthStore } from '../store/authStore';
-import type { LrcLine } from '../api/lrclib';
-import type { Track } from '../store/playerStoreTypes';
-import { EaseScroller, targetForFraction } from '../utils/easeScroll';
 import { usePlaybackDelayPress } from '../hooks/usePlaybackDelayPress';
 import PlaybackDelayModal from './PlaybackDelayModal';
 import PlaybackScheduleBadge from './PlaybackScheduleBadge';
 import { usePlaybackScheduleRemaining } from '../utils/playbackScheduleFormat';
+import { FsLyricsApple } from './fullscreenPlayer/FsLyricsApple';
+import { FsLyricsRail } from './fullscreenPlayer/FsLyricsRail';
 
 function formatTime(seconds: number): string {
   if (!seconds || isNaN(seconds)) return '0:00';
@@ -29,302 +27,6 @@ function formatTime(seconds: number): string {
   const s = Math.floor(seconds % 60);
   return `${m}:${s.toString().padStart(2, '0')}`;
 }
-
-// ─── Apple Music-style fullscreen lyrics ─────────────────────────────────────
-// Full-screen scrollable list. Active line auto-scrolls to ~35% from top.
-// Word-sync runs imperatively (no React re-renders on every time tick).
-// User scroll pauses auto-scroll for 4 s then resumes.
-
-const FsLyricsApple = memo(function FsLyricsApple({ currentTrack }: { currentTrack: Track | null }) {
-  const { syncedLines, wordLines, plainLyrics, loading } = useLyrics(currentTrack);
-  const staticOnly = useAuthStore(s => s.lyricsStaticOnly);
-
-  const useWords = !staticOnly && wordLines !== null && wordLines.length > 0;
-  const lineSrc: LrcLine[] | null = useWords
-    ? (wordLines as WordLyricsLine[]).map(l => ({ time: l.time, text: l.text }))
-    : (syncedLines as LrcLine[] | null);
-  const hasSynced = !staticOnly && lineSrc !== null && lineSrc.length > 0;
-
-  const duration = usePlayerStore(s => s.currentTrack?.duration ?? 0);
-  const seek     = usePlayerStore(s => s.seek);
-
-  const linesRef    = useRef<LrcLine[]>([]);
-  linesRef.current  = hasSynced ? lineSrc! : [];
-
-  // React state only for the active line index — changes are infrequent.
-  const [activeIdx, setActiveIdx]   = useState(-1);
-  const activeIdxRef                = useRef(-1);
-
-  const containerRef  = useRef<HTMLDivElement | null>(null);
-  const scrollerRef   = useRef<EaseScroller | null>(null);
-  const lineRefs      = useRef<(HTMLDivElement | null)[]>([]);
-  const wordRefs      = useRef<HTMLSpanElement[][]>([]);
-  const prevWord      = useRef({ line: -1, word: -1 });
-  const isUserScroll  = useRef(false);
-  const scrollTimer   = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  const setContainerRef = useCallback((el: HTMLDivElement | null) => {
-    containerRef.current = el;
-    scrollerRef.current?.stop();
-    scrollerRef.current = el ? new EaseScroller(el) : null;
-  }, []);
-
-  // Reset everything on track change.
-  useEffect(() => {
-    lineRefs.current   = [];
-    wordRefs.current   = [];
-    prevWord.current   = { line: -1, word: -1 };
-    activeIdxRef.current = -1;
-    setActiveIdx(-1);
-    scrollerRef.current?.jump(0);
-  }, [currentTrack?.id]);
-
-  // Subscribe to playback time — only triggers React setState when line changes.
-  useEffect(() => {
-    if (!hasSynced) return;
-    const apply = (time: number) => {
-      const ls = linesRef.current;
-      if (!ls.length) return;
-      const idx = ls.reduce((acc, line, i) => time >= line.time ? i : acc, -1);
-      if (idx !== activeIdxRef.current) {
-        activeIdxRef.current = idx;
-        setActiveIdx(idx);
-      }
-    };
-    apply(getPlaybackProgressSnapshot().currentTime);
-    return subscribePlaybackProgress(s => apply(s.currentTime));
-  }, [hasSynced, currentTrack?.id]);
-
-  // Ease-scroll active line to ~35% from the top of the container.
-  useEffect(() => {
-    if (activeIdx < 0 || isUserScroll.current) return;
-    const el  = lineRefs.current[activeIdx];
-    const box = containerRef.current;
-    if (!el || !box || !scrollerRef.current) return;
-    scrollerRef.current.scrollTo(targetForFraction(box, el, 0.35));
-  }, [activeIdx]);
-
-  // Word-sync: imperative DOM updates, zero React re-renders per tick.
-  useEffect(() => {
-    wordRefs.current = [];
-    prevWord.current = { line: -1, word: -1 };
-  }, [currentTrack?.id, useWords]);
-
-  useEffect(() => {
-    if (!useWords) return;
-    const lines = wordLines as WordLyricsLine[];
-    const apply = (time: number) => {
-      let li = -1;
-      for (let i = 0; i < lines.length; i++) { if (time >= lines[i].time) li = i; else break; }
-      let wi = -1;
-      if (li >= 0) {
-        const ws = lines[li].words;
-        for (let j = 0; j < ws.length; j++) { if (time >= ws[j].time) wi = j; else break; }
-      }
-      const prev = prevWord.current;
-      if (prev.line === li && prev.word === wi) return;
-      if (prev.line !== li && prev.line >= 0 && wordRefs.current[prev.line])
-        for (const w of wordRefs.current[prev.line]) w.className = 'fsa-lyric-word';
-      if (li >= 0 && wordRefs.current[li]) {
-        const ws = wordRefs.current[li];
-        for (let j = 0; j < ws.length; j++)
-          ws[j].className = j < wi ? 'fsa-lyric-word played' : j === wi ? 'fsa-lyric-word active' : 'fsa-lyric-word';
-      }
-      prevWord.current = { line: li, word: wi };
-    };
-    apply(getPlaybackProgressSnapshot().currentTime);
-    return subscribePlaybackProgress(s => apply(s.currentTime));
-  }, [useWords, wordLines]);
-
-  const handleUserScroll = useCallback(() => {
-    scrollerRef.current?.stop();
-    isUserScroll.current = true;
-    if (scrollTimer.current) clearTimeout(scrollTimer.current);
-    scrollTimer.current = setTimeout(() => { isUserScroll.current = false; }, 4000);
-  }, []);
-
-  const handleClick = useCallback((e: React.MouseEvent<HTMLDivElement>) => {
-    const target = (e.target as HTMLElement).closest<HTMLElement>('[data-time]');
-    if (!target || duration <= 0) return;
-    seek(parseFloat(target.dataset.time!) / duration);
-  }, [duration, seek]);
-
-  if (!currentTrack || loading) return null;
-
-  const isPlain = !hasSynced && !!plainLyrics;
-
-  return (
-    <div
-      className={`fsa-lyrics-container${isPlain ? ' fsa-lyrics-container--plain' : ''}`}
-      ref={setContainerRef}
-      onWheel={handleUserScroll}
-      onTouchMove={handleUserScroll}
-      onClick={handleClick}
-      aria-hidden="true"
-    >
-      <div className="fsa-lyrics-top-pad" />
-
-      {hasSynced && (useWords
-        ? (wordLines as WordLyricsLine[]).map((line, i) => (
-            <div
-              key={i}
-              ref={el => { lineRefs.current[i] = el; }}
-              className={`fsa-lyric-line${i === activeIdx ? ' fsal-active' : i < activeIdx ? ' fsal-past' : ''}`}
-              data-time={line.time}
-            >
-              {line.words.length > 0
-                ? line.words.map((w, j) => (
-                    <span
-                      key={j}
-                      className="fsa-lyric-word"
-                      ref={el => {
-                        if (!wordRefs.current[i]) wordRefs.current[i] = [];
-                        if (el) wordRefs.current[i][j] = el;
-                      }}
-                    >{w.text}</span>
-                  ))
-                : (line.text || '\u00A0')}
-            </div>
-          ))
-        : lineSrc!.map((line, i) => (
-            <div
-              key={i}
-              ref={el => { lineRefs.current[i] = el; }}
-              className={`fsa-lyric-line${i === activeIdx ? ' fsal-active' : i < activeIdx ? ' fsal-past' : ''}`}
-              data-time={line.time}
-            >
-              {line.text || '\u00A0'}
-            </div>
-          ))
-      )}
-
-      {!hasSynced && plainLyrics && (
-        <div className="fsa-plain-lyrics">
-          {plainLyrics.split('\n').map((line, i) => (
-            <p key={i} className="fsa-plain-line">{line || '\u00A0'}</p>
-          ))}
-        </div>
-      )}
-
-      <div className="fsa-lyrics-bottom-pad" />
-    </div>
-  );
-});
-
-// ─── Classic 5-line rail lyrics (original "Rail" style) ──────────────────────
-// Slot height = 6vh = window.innerHeight * 0.06 — must match CSS height: 6vh.
-const FsLyricsRail = memo(function FsLyricsRail({ currentTrack }: { currentTrack: Track | null }) {
-  const { syncedLines, wordLines, loading } = useLyrics(currentTrack);
-  const staticOnly = useAuthStore(s => s.lyricsStaticOnly);
-
-  const useWords  = !staticOnly && wordLines !== null && wordLines.length > 0;
-  const lineSrc: LrcLine[] | null = useWords
-    ? (wordLines as WordLyricsLine[]).map(l => ({ time: l.time, text: l.text }))
-    : (syncedLines as LrcLine[] | null);
-  const hasSynced = !staticOnly && lineSrc !== null && lineSrc.length > 0;
-
-  const linesRef = useRef<LrcLine[]>([]);
-  linesRef.current = hasSynced ? lineSrc! : [];
-
-  const activeIdx = usePlayerStore(s => {
-    const ls = linesRef.current;
-    if (ls.length === 0) return -1;
-    return ls.reduce((acc, line, i) => s.currentTime >= line.time ? i : acc, -1);
-  });
-
-  const duration = usePlayerStore(s => s.currentTrack?.duration ?? 0);
-  const seek     = usePlayerStore(s => s.seek);
-
-  const slotH = useRef(window.innerHeight * 0.06);
-  useEffect(() => {
-    const onResize = () => { slotH.current = window.innerHeight * 0.06; };
-    window.addEventListener('resize', onResize);
-    return () => window.removeEventListener('resize', onResize);
-  }, []);
-
-  const handleLineClick = useCallback((e: React.MouseEvent<HTMLDivElement>) => {
-    const target = (e.target as HTMLElement).closest<HTMLElement>('[data-time]');
-    if (!target || duration <= 0) return;
-    seek(parseFloat(target.dataset.time!) / duration);
-  }, [duration, seek]);
-
-  const wordRefs = useRef<HTMLSpanElement[][]>([]);
-  const prevWord = useRef<{ line: number; word: number }>({ line: -1, word: -1 });
-
-  useEffect(() => {
-    wordRefs.current = [];
-    prevWord.current = { line: -1, word: -1 };
-  }, [currentTrack?.id, useWords]);
-
-  useEffect(() => {
-    if (!useWords) return;
-    const lines = wordLines as WordLyricsLine[];
-    const apply = (time: number) => {
-      let li = -1;
-      for (let i = 0; i < lines.length; i++) { if (time >= lines[i].time) li = i; else break; }
-      let wi = -1;
-      if (li >= 0) {
-        const ws = lines[li].words;
-        for (let j = 0; j < ws.length; j++) { if (time >= ws[j].time) wi = j; else break; }
-      }
-      const prev = prevWord.current;
-      if (prev.line === li && prev.word === wi) return;
-      if (prev.line !== li && prev.line >= 0 && wordRefs.current[prev.line])
-        for (const w of wordRefs.current[prev.line]) w.className = 'fsr-lyric-word';
-      if (li >= 0 && wordRefs.current[li]) {
-        const ws = wordRefs.current[li];
-        for (let j = 0; j < ws.length; j++)
-          ws[j].className = j < wi ? 'fsr-lyric-word played' : j === wi ? 'fsr-lyric-word active' : 'fsr-lyric-word';
-      }
-      prevWord.current = { line: li, word: wi };
-    };
-    apply(getPlaybackProgressSnapshot().currentTime);
-    return subscribePlaybackProgress(s => apply(s.currentTime));
-  }, [useWords, wordLines]);
-
-  if (!currentTrack || loading || !hasSynced) return null;
-
-  const railY = (2 - Math.max(0, activeIdx)) * slotH.current;
-
-  return (
-    <div className="fsr-lyrics-overlay" aria-hidden="true">
-      <div
-        className="fsr-lyrics-rail"
-        style={{ transform: `translateY(${railY}px)` }}
-        onClick={handleLineClick}
-      >
-        {useWords
-          ? (wordLines as WordLyricsLine[]).map((line, i) => (
-              <div
-                key={i}
-                className={`fsr-lyric-line${i === activeIdx ? ' fsrl-active' : i < activeIdx ? ' fsrl-past' : ''}`}
-                data-time={line.time}
-              >
-                {line.words.length > 0 ? line.words.map((w, j) => (
-                  <span
-                    key={j}
-                    className="fsr-lyric-word"
-                    ref={el => {
-                      if (!wordRefs.current[i]) wordRefs.current[i] = [];
-                      if (el) wordRefs.current[i][j] = el;
-                    }}
-                  >{w.text}</span>
-                )) : (line.text || '\u00A0')}
-              </div>
-            ))
-          : lineSrc!.map((line, i) => (
-              <div
-                key={i}
-                className={`fsr-lyric-line${i === activeIdx ? ' fsrl-active' : i < activeIdx ? ' fsrl-past' : ''}`}
-                data-time={line.time}
-              >
-                {line.text || '\u00A0'}
-              </div>
-            ))}
-      </div>
-    </div>
-  );
-});
 
 // ─── Album art box — crossfades layers so old art stays visible while new loads ─
 // Uses 300px thumbnails (portrait fallback uses 500px separately).
